# specifications/architecture/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DigitalBurj Logistics OS API",
    "version": "1.0.0",
    "description": "Versioned REST contract. Money = decimal strings. Commands are idempotent via Idempotency-Key; mutable aggregates use If-Match."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "SERVICE_UNAVAILABLE",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "TENANT_REQUIRED",
              "NOT_FOUND",
              "VALIDATION_FAILED",
              "VERSION_CONFLICT",
              "IDEMPOTENCY_KEY_REUSED",
              "IDEMPOTENCY_KEY_REQUIRED",
              "INVALID_STATE_TRANSITION",
              "DUPLICATE",
              "ACCOUNTING_PERIOD_CLOSED",
              "PERIOD_NOT_FOUND",
              "INSUFFICIENT_STOCK",
              "STOCK_ON_HOLD",
              "RELEASE_EVIDENCE_REQUIRED",
              "SEPARATION_OF_DUTIES",
              "OVER_ALLOCATION",
              "JOB_NOT_CLOSABLE",
              "CURRENCY_MISMATCH",
              "DOCUMENT_NOT_CLEAN",
              "WEBHOOK_SIGNATURE_INVALID",
              "AI_TOOL_DENIED",
              "AI_BUDGET_EXCEEDED",
              "BOOKING_OUTCOME_UNKNOWN",
              "QUALIFICATION_REQUIRED",
              "RATE_LIMITED",
              "INTERNAL"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      }
    },
...
  },
  "paths": {
    "/api/v1/accounting-periods": {
      "get": {
        "operationId": "listAccountingPeriods",
        "tags": [
          "Finance"
        ],
        "summary": "Scoped accounting periods",
        "description": "Permission: finance.period.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounting-periods/{id}": {
      "get": {
        "operationId": "getAccountingPeriod",
        "tags": [
          "Finance"
        ],
        "summary": "Period trial balance and reviewed change history",
        "description": "Permission: finance.period.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounting-periods/{id}/requests": {
      "post": {
        "operationId": "requestPeriodChange",
        "tags": [
          "Finance"
        ],
        "summary": "Request reconciled period close or controlled reopening",
        "description": "Permission: finance.period.request",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "close",
                      "reopen"
                    ]
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 4000
                  }
                },
                "required": [
                  "kind",
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/period-requests/{id}/decide": {
      "post": {
        "operationId": "decidePeriodChange",
        "tags": [
          "Finance"
        ],
        "summary": "Independently approve or reject a period change",
        "description": "Permission: finance.period.decide",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "approve": {
                    "type": "boolean"
                  },
                  "note": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 4000
                  }
                },
                "required": [
                  "approve",
                  "note"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rfqs": {
      "get": {
        "operationId": "listRfqs",
        "tags": [
          "Procurement"
        ],
        "summary": "Legal-entity scoped carrier rate requests",
        "description": "Permission: rates.view",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createRfq",
        "tags": [
          "Procurement"
        ],
        "summary": "Prepare a carrier rate request with invited suppliers",
        "description": "Permission: rates.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalEntityId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "ocean_fcl",
                      "ocean_lcl",
                      "air",
                      "road",
                      "multimodal",
                      "warehouse",
                      "customs"
                    ]
                  },
                  "origin": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "destination": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "requirements": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 10000
                  },
                  "currency": {
                    "type": "string",
                    "pattern": "^[A-Z]{3}$"
                  },
                  "responseDeadline": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "supplierPartyIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "minItems": 1,
                    "maxItems": 50
                  }
                },
                "required": [
                  "legalEntityId",
                  "mode",
                  "origin",
                  "destination",
                  "requirements",
                  "currency",
                  "responseDeadline",
                  "supplierPartyIds"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rfqs/{id}/issue": {
      "post": {
        "operationId": "issueRfq",
        "tags": [
          "Procurement"
        ],
        "summary": "Issue and freeze a request for manual supplier distribution",
        "description": "Permission: rates.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rfqs/{id}/offers": {
      "post": {
        "operationId": "recordRfqOffer",
        "tags": [
          "Procurement"
        ],
        "summary": "Record an invited supplier response as an immutable revision",
        "description": "Permission: rates.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "supplierPartyId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "freight": {
                    "type": "string",
                    "pattern": "^-?\\d+(\\.\\d{1,4})?$"
                  },
                  "localCharges": {
                    "type": "string",
                    "pattern": "^-?\\d+(\\.\\d{1,4})?$"
                  },
                  "transitDays": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 3650
                  },
                  "freeDays": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 3650
                  },
                  "validUntil": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "terms": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 10000
                  }
                },
                "required": [
                  "supplierPartyId",
                  "freight",
                  "localCharges",
                  "transitDays",
                  "freeDays",
                  "validUntil",
                  "terms"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rfqs/{id}/award": {
      "post": {
        "operationId": "awardRfq",
        "tags": [
          "Procurement"
        ],
        "summary": "Independently award a valid latest offer and publish an approved rate",
        "description": "Permission: rates.manage",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "X-Tenant-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "offerId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 4000
                  }
                },
                "required": [
                  "offerId",
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "4XX": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}